Combine actions registry and resolvers store
