+++ lib/clowne/plan.rb
@@ -2,61 +2,7 @@
 
 module Clowne
   class Plan # :nodoc: all
-    class Registry
-      attr_reader :actions
-
-      def initialize
-        @actions = []
-      end
-
-      def insert_after(after, action)
-        validate_uniq!(action)
-
-        after_index = actions.find_index(after)
-
-        raise "Plan action not found: #{after}" if after_index.nil?
-
-        actions.insert(after_index + 1, action)
-      end
-
-      def insert_before(before, action)
-        validate_uniq!(action)
-
-        before_index = actions.find_index(before)
-
-        raise "Plan action not found: #{before}" if before_index.nil?
-
-        actions.insert(before_index, action)
-      end
-
-      def append(action)
-        validate_uniq!(action)
-        actions.push action
-      end
-
-      def prepend(action)
-        validate_uniq!(action)
-        actions.unshift action
-      end
-
-      private
-
-      def validate_uniq!(action)
-        raise "Plan action already registered: #{action}" if actions.include?(action)
-      end
-    end
-
-    class << self
-      attr_reader :registry
-
-      protected
-
-      attr_writer :registry
-    end
-
-    self.registry = Registry.new
-
-    def initialize(registry = self.class.registry)
+    def initialize(registry)
       @registry = registry
       @data = {}
     end
