+++ spec/clowne/cloner_spec.rb
@@ -1,7 +1,7 @@
 describe Clowne::Cloner do
   let(:cloner) do
     Class.new(described_class) do
-      adapter FakeAdapter
+      adapter :active_record
 
       include_all
 
@@ -91,7 +91,7 @@ describe Clowne::Cloner do
     context 'when object is nil' do
       let(:cloner) do
         Class.new(Clowne::Cloner) do
-          adapter FakeAdapter
+          adapter :base
         end
       end
 
@@ -106,7 +106,7 @@ describe Clowne::Cloner do
     context 'when trait is unknown' do
       let(:cloner) do
         Class.new(Clowne::Cloner) do
-          adapter FakeAdapter
+          adapter :active_record
 
           trait :with_comments do
             include_association :comments
