--- spec/support/fake_adapter.rb
@@ -1,2 +1,0 @@
-class FakeAdapter < Clowne::Adapters::Base
-end
